Version 1.0
Automatically display the first two pictures after selecting a folder.
Before it was necessary to click on one of the buttons to get started;
that's not user friendly.
Cleaned up the print statements.
Renaming of .prsettings because Windows doesn't like files without
a name and only an extension... Plus, on Linux .prsettings is hidden,
which is annoying because there are other hidden files created by
git and eclipse. In short, the new name works better.
README lists some further TODOs and stuff on my wish list.